requirements.txt
py-clob-client
python-dotenv
requests
web3
pytest
fastapi
hypercorn
boto3
pydantic
py-cord
aiohttp
redis
websockets
